requirements.txt
flask==3.0.0
gunicorn==21.2.0
google-cloud-vision==3.7.2